requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2
django-bootstrap5==24.3
django-celery-beat==2.8.0
django-crispy-forms==2.3
django-timezone-field==7.1
django-user-agents==0.4.0
djangorestframework==3.16.0
idna==3.10
kombu==5.5.3
pillow==11.0.0
prompt_toolkit==3.0.51
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
